+++ public/lib/maxmind/GeoIp2/src/Model/Isp.php
@@ -71,8 +71,8 @@ class Isp implements \JsonSerializable
     public function __construct(array $raw)
     {
         $this->autonomousSystemNumber = $raw['autonomous_system_number'] ?? null;
-        $this->autonomousSystemOrganization =
-            $raw['autonomous_system_organization'] ?? null;
+        $this->autonomousSystemOrganization
+            = $raw['autonomous_system_organization'] ?? null;
         $this->isp = $raw['isp'] ?? null;
         $this->mobileCountryCode = $raw['mobile_country_code'] ?? null;
         $this->mobileNetworkCode = $raw['mobile_network_code'] ?? null;